requirements.txt
Django==4.0.4
environs==9.5.0
Pillow==9.2.0
requests==2.27.1
django-crispy-forms==1.14.0
psycopg2-binary==2.9.5
django-object-actions==4.0.0
djangorestframework==3.14.0
drf-yasg==1.21.4
requests-toolbelt==0.10.1